% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcustom-package.R
\docType{package}
\name{Rcustom-package}
\alias{Rcustom}
\alias{Rcustom-package}
\title{Rcustom: Custom code}
\description{
Custom code
}
\author{
\strong{Maintainer}: Matias Faure \email{hi@faure.hu}

}
\keyword{internal}
